Refactor MainAgentPage to use custom hook useMainAgentPage
